lab4/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==1.4.0
alembic==1.13.1
aiosqlite==0.19.0
pydantic==1.10.13
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0
